add controller and controller test
