+++ src/main/java/com/tenniscourts/dto/CreateReservationRequestDTO.java
@@ -1,0 +1,26 @@
+package com.tenniscourts.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import javax.validation.constraints.NotNull;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+@Setter
+@Builder
+@Data
+public class CreateReservationRequestDTO {
+
+    @NotNull
+    private Long guestId;
+
+    @NotNull
+    private Long scheduleId;
+
+}
