+++ src/main/java/com/tenniscourts/dto/CreateScheduleRequestDTO.java
@@ -1,0 +1,21 @@
+package com.tenniscourts.dto;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.validation.constraints.NotNull;
+import java.time.LocalDateTime;
+
+@Getter
+@Setter
+public class CreateScheduleRequestDTO {
+
+    @NotNull
+    private Long tennisCourtId;
+
+    @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm")
+    @NotNull
+    private LocalDateTime startDateTime;
+
+}
