+++ src/main/java/com/tenniscourts/dto/ScheduleDTO.java
@@ -1,0 +1,28 @@
+package com.tenniscourts.dto;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.validation.constraints.NotNull;
+import java.time.LocalDateTime;
+
+@Getter
+@Setter
+public class ScheduleDTO {
+
+    private Long id;
+
+    private TennisCourtDTO tennisCourt;
+
+    @NotNull
+    private Long tennisCourtId;
+
+    @JsonFormat(pattern="yyyy-MM-dd'T'HH:mm")
+    @NotNull
+    private LocalDateTime startDateTime;
+
+    @JsonFormat(pattern="yyyy-MM-dd'T'HH:mm")
+    private LocalDateTime endDateTime;
+
+}
