+++ src/main/java/com/tenniscourts/dto/TennisCourtDTO.java
@@ -1,0 +1,29 @@
+package com.tenniscourts.dto;
+
+import com.tenniscourts.dto.ScheduleDTO;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+
+import javax.validation.constraints.NotNull;
+import java.util.List;
+
+@Getter
+@Setter
+@ToString
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class TennisCourtDTO {
+
+    private Long id;
+
+    @NotNull
+    private String name;
+
+    private List<ScheduleDTO> tennisCourtSchedules;
+
+}
