+++ src/main/java/com/tenniscourts/mapper/ReservationMapper.java
@@ -1,0 +1,22 @@
+package com.tenniscourts.mapper;
+
+import com.tenniscourts.dto.CreateReservationRequestDTO;
+import com.tenniscourts.dto.ReservationDTO;
+import com.tenniscourts.model.Reservation;
+import org.mapstruct.InheritInverseConfiguration;
+import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
+import org.mapstruct.ReportingPolicy;
+
+@Mapper(componentModel = "spring", unmappedTargetPolicy = ReportingPolicy.IGNORE)
+public interface ReservationMapper {
+
+    Reservation map(ReservationDTO source);
+
+    @InheritInverseConfiguration
+    ReservationDTO map(Reservation source);
+
+    @Mapping(target = "guest.id", source = "guestId")
+    @Mapping(target = "schedule.id", source = "scheduleId")
+    Reservation map(CreateReservationRequestDTO source);
+}
