+++ src/main/java/com/tenniscourts/mapper/TennisCourtMapper.java
@@ -1,0 +1,15 @@
+package com.tenniscourts.mapper;
+
+import com.tenniscourts.dto.TennisCourtDTO;
+import com.tenniscourts.model.TennisCourt;
+import org.mapstruct.InheritInverseConfiguration;
+import org.mapstruct.Mapper;
+import org.mapstruct.ReportingPolicy;
+
+@Mapper(componentModel = "spring", unmappedTargetPolicy = ReportingPolicy.IGNORE)
+public interface TennisCourtMapper {
+    TennisCourtDTO map(TennisCourt source);
+
+    @InheritInverseConfiguration
+    TennisCourt map(TennisCourtDTO source);
+}
