+++ src/main/java/com/tenniscourts/model/Guest.java
@@ -1,0 +1,32 @@
+package com.tenniscourts.model;
+
+import com.tenniscourts.config.persistence.BaseEntity;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Table;
+import javax.validation.constraints.NotNull;
+
+@Entity
+@Table
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@EqualsAndHashCode(callSuper = true)
+@ToString
+public class Guest extends BaseEntity<Long> {
+
+  @Column
+  @NotNull
+  private String name;
+
+}
