+++ src/main/java/com/tenniscourts/model/ReservationStatus.java
@@ -1,0 +1,7 @@
+package com.tenniscourts.model;
+
+public enum ReservationStatus {
+  READY_TO_PLAY,
+  CANCELLED,
+  RESCHEDULED
+}
