+++ src/main/java/com/tenniscourts/model/Schedule.java
@@ -1,0 +1,54 @@
+package com.tenniscourts.model;
+
+import com.tenniscourts.config.persistence.BaseEntity;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.ManyToOne;
+import javax.persistence.OneToMany;
+import javax.persistence.Table;
+import javax.validation.constraints.NotNull;
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+@Table
+@Getter
+@Setter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+@EqualsAndHashCode(callSuper = true, exclude = "reservations")
+public class Schedule extends BaseEntity<Long> {
+
+    @ManyToOne
+    @NotNull
+    private TennisCourt tennisCourt;
+
+    @Column
+    @NotNull
+    private LocalDateTime startDateTime;
+
+    @Column
+    @NotNull
+    private LocalDateTime endDateTime;
+
+    @OneToMany
+    private List<Reservation> reservations;
+
+    public void addReservation(Reservation reservation) {
+        if (this.reservations == null) {
+            this.reservations = new ArrayList<>();
+        }
+
+        reservation.setSchedule(this);
+        this.reservations.add(reservation);
+    }
+}
