+++ src/main/java/com/tenniscourts/model/TennisCourt.java
@@ -1,0 +1,29 @@
+package com.tenniscourts.model;
+
+import com.tenniscourts.config.persistence.BaseEntity;
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Table;
+import javax.validation.constraints.NotNull;
+
+@Entity
+@Table
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@EqualsAndHashCode(callSuper = true)
+@ToString
+public class TennisCourt extends BaseEntity<Long> {
+
+    @Column
+    @NotNull
+    private String name;
+}
