+++ src/main/java/com/tenniscourts/repository/ScheduleRepository.java
@@ -1,0 +1,11 @@
+package com.tenniscourts.repository;
+
+import com.tenniscourts.model.Schedule;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface ScheduleRepository extends JpaRepository<Schedule, Long> {
+
+    List<Schedule> findByTennisCourt_IdOrderByStartDateTime(Long id);
+}
