+++ src/main/java/com/tenniscourts/repository/TennisCourtRepository.java
@@ -1,0 +1,7 @@
+package com.tenniscourts.repository;
+
+import com.tenniscourts.model.TennisCourt;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TennisCourtRepository extends JpaRepository<TennisCourt, Long> {
+}
