--- src/main/java/com/tenniscourts/reservations/ReservationController.java
@@ -1,27 +1,0 @@
-package com.tenniscourts.reservations;
-
-import com.tenniscourts.config.BaseRestController;
-import lombok.AllArgsConstructor;
-import org.springframework.http.ResponseEntity;
-
-@AllArgsConstructor
-public class ReservationController extends BaseRestController {
-
-    private final ReservationService reservationService;
-
-    public ResponseEntity<Void> bookReservation(CreateReservationRequestDTO createReservationRequestDTO) {
-        return ResponseEntity.created(locationByEntity(reservationService.bookReservation(createReservationRequestDTO).getId())).build();
-    }
-
-    public ResponseEntity<ReservationDTO> findReservation(Long reservationId) {
-        return ResponseEntity.ok(reservationService.findReservation(reservationId));
-    }
-
-    public ResponseEntity<ReservationDTO> cancelReservation(Long reservationId) {
-        return ResponseEntity.ok(reservationService.cancelReservation(reservationId));
-    }
-
-    public ResponseEntity<ReservationDTO> rescheduleReservation(Long reservationId, Long scheduleId) {
-        return ResponseEntity.ok(reservationService.rescheduleReservation(reservationId, scheduleId));
-    }
-}
