--- src/main/java/com/tenniscourts/reservations/ReservationMapper.java
@@ -1,18 +1,0 @@
-package com.tenniscourts.reservations;
-
-import org.mapstruct.InheritInverseConfiguration;
-import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
-
-@Mapper(componentModel = "spring")
-public interface ReservationMapper {
-
-    Reservation map(ReservationDTO source);
-
-    @InheritInverseConfiguration
-    ReservationDTO map(Reservation source);
-
-    @Mapping(target = "guest.id", source = "guestId")
-    @Mapping(target = "schedule.id", source = "scheduleId")
-    Reservation map(CreateReservationRequestDTO source);
-}
