--- src/main/java/com/tenniscourts/reservations/ReservationRepository.java
@@ -1,15 +1,0 @@
-package com.tenniscourts.reservations;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-public interface ReservationRepository extends JpaRepository<Reservation, Long> {
-
-    List<Reservation> findBySchedule_Id(Long scheduleId);
-
-    List<Reservation> findByReservationStatusAndSchedule_StartDateTimeGreaterThanEqualAndSchedule_EndDateTimeLessThanEqual(ReservationStatus reservationStatus, LocalDateTime startDateTime, LocalDateTime endDateTime);
-
-//    List<Reservation> findByStartDateTimeGreaterThanEqualAndEndDateTimeLessThanEqualAndTennisCourt(LocalDateTime startDateTime, LocalDateTime endDateTime, TennisCourt tennisCourt);
-}
