--- src/main/java/com/tenniscourts/reservations/ReservationStatus.java
@@ -1,7 +1,0 @@
-package com.tenniscourts.reservations;
-
-public enum ReservationStatus {
-  READY_TO_PLAY,
-  CANCELLED,
-  RESCHEDULED
-}
