--- src/main/java/com/tenniscourts/schedules/CreateScheduleRequestDTO.java
@@ -1,21 +1,0 @@
-package com.tenniscourts.schedules;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.validation.constraints.NotNull;
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-public class CreateScheduleRequestDTO {
-
-    @NotNull
-    private Long tennisCourtId;
-
-    @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm")
-    @NotNull
-    private LocalDateTime startDateTime;
-
-}
