--- src/main/java/com/tenniscourts/schedules/ScheduleDTO.java
@@ -1,29 +1,0 @@
-package com.tenniscourts.schedules;
-
-import com.tenniscourts.tenniscourts.TennisCourtDTO;
-import com.fasterxml.jackson.annotation.JsonFormat;
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.validation.constraints.NotNull;
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-public class ScheduleDTO {
-
-    private Long id;
-
-    private TennisCourtDTO tennisCourt;
-
-    @NotNull
-    private Long tennisCourtId;
-
-    @JsonFormat(pattern="yyyy-MM-dd'T'HH:mm")
-    @NotNull
-    private LocalDateTime startDateTime;
-
-    @JsonFormat(pattern="yyyy-MM-dd'T'HH:mm")
-    private LocalDateTime endDateTime;
-
-}
