--- src/main/java/com/tenniscourts/schedules/ScheduleRepository.java
@@ -1,10 +1,0 @@
-package com.tenniscourts.schedules;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface ScheduleRepository extends JpaRepository<Schedule, Long> {
-
-    List<Schedule> findByTennisCourt_IdOrderByStartDateTime(Long id);
-}
