--- src/main/java/com/tenniscourts/schedules/ScheduleService.java
@@ -1,35 +1,0 @@
-package com.tenniscourts.schedules;
-
-import lombok.AllArgsConstructor;
-import org.springframework.stereotype.Service;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-@Service
-@AllArgsConstructor
-public class ScheduleService {
-
-    private final ScheduleRepository scheduleRepository;
-
-    private final ScheduleMapper scheduleMapper;
-
-    public ScheduleDTO addSchedule(Long tennisCourtId, CreateScheduleRequestDTO createScheduleRequestDTO) {
-        //TODO: implement addSchedule
-        return null;
-    }
-
-    public List<ScheduleDTO> findSchedulesByDates(LocalDateTime startDate, LocalDateTime endDate) {
-        //TODO: implement
-        return null;
-    }
-
-    public ScheduleDTO findSchedule(Long scheduleId) {
-        //TODO: implement
-        return null;
-    }
-
-    public List<ScheduleDTO> findSchedulesByTennisCourtId(Long tennisCourtId) {
-        return scheduleMapper.map(scheduleRepository.findByTennisCourt_IdOrderByStartDateTime(tennisCourtId));
-    }
-}
