+++ src/main/java/com/tenniscourts/service/ReservationService.java
@@ -1,0 +1,99 @@
+package com.tenniscourts.service;
+
+import com.tenniscourts.dto.CreateReservationRequestDTO;
+import com.tenniscourts.dto.ReservationDTO;
+import com.tenniscourts.exceptions.EntityNotFoundException;
+import com.tenniscourts.mapper.ReservationMapper;
+import com.tenniscourts.model.Reservation;
+import com.tenniscourts.repository.ReservationRepository;
+import com.tenniscourts.model.ReservationStatus;
+import lombok.AllArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
+import java.time.temporal.ChronoUnit;
+
+@Service
+@AllArgsConstructor
+public class ReservationService {
+
+    private final ReservationRepository reservationRepository;
+
+    private final ReservationMapper reservationMapper;
+
+    public ReservationDTO bookReservation(CreateReservationRequestDTO createReservationRequestDTO) {
+        throw new UnsupportedOperationException();
+    }
+
+    public ReservationDTO findReservation(Long reservationId) {
+        return reservationRepository.findById(reservationId).map(reservationMapper::map).<EntityNotFoundException>orElseThrow(() -> {
+            throw new EntityNotFoundException("Reservation not found.");
+        });
+    }
+
+    public ReservationDTO cancelReservation(Long reservationId) {
+        return reservationMapper.map(this.cancel(reservationId));
+    }
+
+    private Reservation cancel(Long reservationId) {
+        return reservationRepository.findById(reservationId).map(reservation -> {
+
+            this.validateCancellation(reservation);
+
+            BigDecimal refundValue = getRefundValue(reservation);
+            return this.updateReservation(reservation, refundValue, ReservationStatus.CANCELLED);
+
+        }).<EntityNotFoundException>orElseThrow(() -> {
+            throw new EntityNotFoundException("Reservation not found.");
+        });
+    }
+
+    private Reservation updateReservation(Reservation reservation, BigDecimal refundValue, ReservationStatus status) {
+        reservation.setReservationStatus(status);
+        reservation.setValue(reservation.getValue().subtract(refundValue));
+        reservation.setRefundValue(refundValue);
+
+        return reservationRepository.save(reservation);
+    }
+
+    private void validateCancellation(Reservation reservation) {
+        if (!ReservationStatus.READY_TO_PLAY.equals(reservation.getReservationStatus())) {
+            throw new IllegalArgumentException("Cannot cancel/reschedule because it's not in ready to play status.");
+        }
+
+        if (reservation.getSchedule().getStartDateTime().isBefore(LocalDateTime.now())) {
+            throw new IllegalArgumentException("Can cancel/reschedule only future dates.");
+        }
+    }
+
+    public BigDecimal getRefundValue(Reservation reservation) {
+        long hours = ChronoUnit.HOURS.between(LocalDateTime.now(), reservation.getSchedule().getStartDateTime());
+
+        if (hours >= 24) {
+            return reservation.getValue();
+        }
+
+        return BigDecimal.ZERO;
+    }
+
+    /*TODO: This method actually not fully working, find a way to fix the issue when it's throwing the error:
+            "Cannot reschedule to the same slot.*/
+    public ReservationDTO rescheduleReservation(Long previousReservationId, Long scheduleId) {
+        Reservation previousReservation = cancel(previousReservationId);
+
+        if (scheduleId.equals(previousReservation.getSchedule().getId())) {
+            throw new IllegalArgumentException("Cannot reschedule to the same slot.");
+        }
+
+        previousReservation.setReservationStatus(ReservationStatus.RESCHEDULED);
+        reservationRepository.save(previousReservation);
+
+        ReservationDTO newReservation = bookReservation(CreateReservationRequestDTO.builder()
+                .guestId(previousReservation.getGuest().getId())
+                .scheduleId(scheduleId)
+                .build());
+        newReservation.setPreviousReservation(reservationMapper.map(previousReservation));
+        return newReservation;
+    }
+}
