+++ src/main/java/com/tenniscourts/service/ScheduleService.java
@@ -1,0 +1,43 @@
+package com.tenniscourts.service;
+
+import com.tenniscourts.dto.CreateScheduleRequestDTO;
+import com.tenniscourts.dto.ScheduleDTO;
+import com.tenniscourts.mapper.ScheduleMapper;
+import com.tenniscourts.repository.ScheduleRepository;
+import lombok.AllArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+@Service
+public class ScheduleService {
+
+    private final ScheduleRepository scheduleRepository;
+
+    private final ScheduleMapper scheduleMapper;
+
+    public ScheduleService(final ScheduleRepository scheduleRepository, final ScheduleMapper scheduleMapper){
+        this.scheduleRepository = scheduleRepository;
+        this.scheduleMapper = scheduleMapper;
+    }
+
+    public ScheduleDTO addSchedule(Long tennisCourtId, CreateScheduleRequestDTO createScheduleRequestDTO) {
+        //TODO: implement addSchedule
+        return new ScheduleDTO();
+    }
+
+    public List<ScheduleDTO> findSchedulesByDates(LocalDateTime startDate, LocalDateTime endDate) {
+        //TODO: implement
+        return null;
+    }
+
+    public ScheduleDTO findSchedule(Long scheduleId) {
+        //TODO: implement
+        return null;
+    }
+
+    public List<ScheduleDTO> findSchedulesByTennisCourtId(Long tennisCourtId) {
+        return scheduleMapper.map(scheduleRepository.findByTennisCourt_IdOrderByStartDateTime(tennisCourtId));
+    }
+}
