--- src/main/java/com/tenniscourts/tenniscourts/TennisCourtController.java
@@ -1,26 +1,0 @@
-package com.tenniscourts.tenniscourts;
-
-import com.tenniscourts.config.BaseRestController;
-import lombok.AllArgsConstructor;
-import org.springframework.http.ResponseEntity;
-
-@AllArgsConstructor
-public class TennisCourtController extends BaseRestController {
-
-    private final TennisCourtService tennisCourtService;
-
-    //TODO: implement rest and swagger
-    public ResponseEntity<Void> addTennisCourt(TennisCourtDTO tennisCourtDTO) {
-        return ResponseEntity.created(locationByEntity(tennisCourtService.addTennisCourt(tennisCourtDTO).getId())).build();
-    }
-
-    //TODO: implement rest and swagger
-    public ResponseEntity<TennisCourtDTO> findTennisCourtById(Long tennisCourtId) {
-        return ResponseEntity.ok(tennisCourtService.findTennisCourtById(tennisCourtId));
-    }
-
-    //TODO: implement rest and swagger
-    public ResponseEntity<TennisCourtDTO> findTennisCourtWithSchedulesById(Long tennisCourtId) {
-        return ResponseEntity.ok(tennisCourtService.findTennisCourtWithSchedulesById(tennisCourtId));
-    }
-}
