--- src/main/java/com/tenniscourts/tenniscourts/TennisCourtMapper.java
@@ -1,12 +1,0 @@
-package com.tenniscourts.tenniscourts;
-
-import org.mapstruct.InheritInverseConfiguration;
-import org.mapstruct.Mapper;
-
-@Mapper(componentModel = "spring")
-public interface TennisCourtMapper {
-    TennisCourtDTO map(TennisCourt source);
-
-    @InheritInverseConfiguration
-    TennisCourt map(TennisCourtDTO source);
-}
