--- src/main/java/com/tenniscourts/tenniscourts/TennisCourtRepository.java
@@ -1,6 +1,0 @@
-package com.tenniscourts.tenniscourts;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TennisCourtRepository extends JpaRepository<TennisCourt, Long> {
-}
