+++ src/test/java/com/tenniscourts/config/WebTestConfig.java
@@ -1,0 +1,15 @@
+package com.tenniscourts.config;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+
+@Configuration
+@EnableWebMvc
+@ComponentScan(basePackages = {"com.tenniscourts.controller"} )
+public class WebTestConfig {
+
+
+}
