+++ src/test/java/com/tenniscourts/service/ScheduleServiceTest.java
@@ -1,0 +1,32 @@
+package com.tenniscourts.service;
+
+import com.tenniscourts.model.Reservation;
+import com.tenniscourts.model.Schedule;
+import org.junit.Assert;
+import org.junit.FixMethodOrder;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.MethodSorters;
+import org.mockito.InjectMocks;
+import org.mockito.junit.MockitoJUnitRunner;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ContextConfiguration;
+
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
+
+@FixMethodOrder(MethodSorters.NAME_ASCENDING)
+@SpringBootTest
+@RunWith(MockitoJUnitRunner.class)
+@ContextConfiguration(classes = ScheduleService.class)
+public class ScheduleServiceTest {
+
+    @InjectMocks
+    ScheduleService scheduleService;
+
+    @Test
+    public void addSchedule() {
+         }
+
+
+}
